feat(terraform): add variables for creating resource group, location, and resource group name.
This commit adds three variables to the Terraform configuration file. The first variable is a boolean that determines whether or not to create a new resource group. The second variable is a string that specifies the location of the resource group. The third variable is a string that specifies the name of the resource group.